updated internal geometry format to support convex triangle fans (beware map version bump)
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@3715 07af1644-1164-4067-b897-63bf4dc9b748

+++ engine/water.cpp
@@ -174,7 +174,7 @@ void rendervertwater(uint subdiv, int xo, int yo, int z, uint size, uchar mat)
     {
         case MAT_WATER:
         {
-            whoffset = fmod(lastmillis/(renderpath!=R_FIXEDFUNCTION ? 600.0f : 300.0f)/(2*M_PI), 1.0f);
+            whoffset = fmod(float(lastmillis/(renderpath!=R_FIXEDFUNCTION ? 600.0f : 300.0f)/(2*M_PI)), 1.0f);
             if(renderpath!=R_FIXEDFUNCTION) { renderwaterstrips(vertwt, z); }
             else 
             {
@@ -195,7 +195,7 @@ void rendervertwater(uint subdiv, int xo, int yo, int z, uint size, uchar mat)
 
         case MAT_LAVA:
         {
-            whoffset = fmod(lastmillis/2000.0f/(2*M_PI), 1.0f);
+            whoffset = fmod(float(lastmillis/2000.0f/(2*M_PI)), 1.0f);
             renderwaterstrips(vertl, z);
             break;
         }
